+++ lib/hlalib.py
@@ -164,7 +164,7 @@ def _levenshtein_distance(first, second):
     return distance_matrix[first_length-1][second_length-1]
 
 
-def getElements(group, return_list=False):
+def getElements(group, return_list=False, include_virtual=False):
     """
     return list of elements.
 
@@ -177,10 +177,14 @@ def getElements(group, return_list=False):
       >>> getElements('F*G1C0*')
       >>> getElements(['FH2G1C30A', 'FH2G1C28A'])
 
-    this calls :func:`~hla.lattice.Lattice.getElements` of the current lattice.
+    this calls :func:`~aphla.lattice.Lattice.getElements` of the current
+    lattice.
     """
 
-    return machines._lat.getElements(group, return_list=return_list)
+    ret = machines._lat.getElements(group, return_list=return_list)
+    if include_virtual == True: return ret
+    else:
+        return [elem for elem in ret if elem.virtual == 0]
 
 def getLocations(group):
     """