+++ doc/history.rst
@@ -6,6 +6,8 @@ v0.3.0
 
   - add pickle/shelve support for :mod:`~hla.element.Element`
   - fixed testing case for element
+  - fix bug in correctOrbitPV
+  - fix bug in getLocations to ignore virtual element
 
 v0.2.2
 ------